Use nsIMsgFolderListener for some notifications
Change to use nsIMsgFolderListener for add/remove notifications.
This way we don't get OnItemAdded/OnItemRemoved spam when moving
a bunch of messages between folders. In the nsIMsgFolderListener
parlance, that would trigger a msgsMoveCopyCompleted instead; we
can just ignore that, because that won't change the net total of
unread messages.

Also in msgAdded/msgsDeleted handlers, check to see that the msg
in question is unread. If it isn't, then that won't change total
unread count either.